pyproject.toml
[tool.poetry]
name = "ads-data-hub"
version = "0.1.0"
description = "Ads Data Hub API client for Python"
authors = ["yassun7010 <yassun7010@outlook.com>"]
readme = "README.md"
packages = [{ include = "ads_data_hub", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.114.0"
pydantic = "^2.6.1"
typing-extensions = "^4.9.0"


[tool.poetry.group.dev.dependencies]
google-api-python-client-stubs = "^1.23.0"
pytest = "^7.4.4"
taskipy = "^1.12.2"
pyright = "^1.1.348"
ipython = "^8.20.0"
ruff = "^0.1.14"

[tool.taskipy.tasks]
format = "ruff format ."
lint = "ruff check ."
typecheck = "pyright src/** tests/** scripts/**"
test = "pytest"
ci = "task format && task lint && task typecheck && task test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
